refactor: clean data declaration
