--- jest.config.js
@@ -1,6 +1,0 @@
-'use strict'
-module.exports = {
-  transform: {
-    '^.+\\.tsx?$': 'ts-jest'
-  }
-}
