--- src/index.d.ts
@@ -1,5 +1,0 @@
-export type Token = string
-
-export interface Tokens {
-  [key: string]: Token
-}
