+++ src/lib/MediaQuery.d.ts
