+++ src/lib/mediaQuery.ts
@@ -1,5 +1,17 @@
 import { MediaQuery } from './mediaQuery.d'
 
+export const findIndexOfMediaQueries = (css: string): number[] => {
+  const indexList = []
+  const len = '@media'.length
+  let currentIndex = css.indexOf('@media')
+
+  while (currentIndex >= 0) {
+    indexList.push(currentIndex)
+    currentIndex = css.indexOf('@media', currentIndex + len)
+  }
+  return indexList
+}
+
 export const getMediaQueries = (css: string): MediaQuery[] => {
   const indexList = findIndexOfMediaQueries(css)
   const mediaQueries: MediaQuery[] = []
@@ -11,27 +23,19 @@ export const getMediaQueries = (css: string): MediaQuery[] => {
     let i = firstParenthesis + 1
     if (i > 0) {
       for (i; deep > 0 && i <= css.length; i++) {
-        if (css[i] === '{') {
-          deep++
-        } else if (css[i] === '}') {
-          deep--
+        if (css[i] === '/' && css[i + 1] === '*') {
+          continue
+        } else {
+          if (css[i] === '{') {
+            deep++
+          } else if (css[i] === '}') {
+            deep--
+          }
+          val += css[i]
         }
-        val += css[i]
       }
       mediaQueries.push({ rule, val, start: index, end: i })
     }
   }
   return mediaQueries
-}
-
-export const findIndexOfMediaQueries = (css: string): number[] => {
-  const indexList = []
-  const len = '@media'.length
-  let currentIndex = css.indexOf('@media')
-
-  while (currentIndex >= 0) {
-    indexList.push(currentIndex)
-    currentIndex = css.indexOf('@media', currentIndex + len)
-  }
-  return indexList
 }
